requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pymongo==4.7.0
python-dotenv==1.0.0
secure==0.3.0
slowapi==0.1.5
bleach==6.1.0
